django_job_search/requirements.txt
django[argon2]==4.2.7
djangorestframework==3.14.0
django-jsonform==2.21.2
django-redis==5.4.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
psycopg2==2.9.9
django-silk==5.0.4
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5